fix(backend): address review feedback on llamacpp wrapper
Split llamacpp.go into llamacpp.go (exported Ping) and
llamacpp_internal.go (unexported newLlamaCpp constructor),
mirroring the vllm.go / vllm_internal.go split that the
TestNoMixedVisibility audit enforces.

Register llamacppFactory in factories_internal.go and add the
{NameLlamaCpp, llamacppFactory} pair to RegisterAll in
factories.go so Registry.Resolve("llamacpp") works.

Add the SPDX license header to llamacpp_e2e_test.go.
Drop the dead return after t.Fatalf in TestNewLlamaCpp_DefaultName.

Addresses: https://github.com/ActiveMemory/ctx/pull/97#issuecomment-4601769149

Spec: specs/ctx-ai-backend.md
