+++ internal/backend/llamacpp_e2e_test.go
@@ -1,3 +1,9 @@
+//   /    ctx:                         https://ctx.ist
+// ,'`./    do you remember?
+// `.,'\
+//   \    Copyright 2026-present Context contributors.
+//                 SPDX-License-Identifier: Apache-2.0
+
 //go:build e2e
 
 package backend
